+++ terraform/examples/slurm_cluster/hybrid/simple/main.tf
@@ -46,6 +46,7 @@ locals {
           enable_confidential_vm   = false
           enable_oslogin           = true
           enable_shielded_vm       = false
+          enable_spot_vm           = false
           gpu                      = null
           instance_template        = null
           labels                   = {}
@@ -56,6 +57,7 @@ locals {
           preemptible              = false
           service_account          = module.slurm_sa_iam["compute"].service_account
           shielded_instance_config = null
+          spot_instance_config     = null
           source_image_family      = null
           source_image_project     = null
           source_image             = null
@@ -94,6 +96,7 @@ locals {
           enable_confidential_vm = false
           enable_oslogin         = true
           enable_shielded_vm     = false
+          enable_spot_vm         = false
           gpu = {
             count = 1
             type  = "nvidia-tesla-v100"
@@ -107,6 +110,7 @@ locals {
           preemptible              = false
           service_account          = module.slurm_sa_iam["compute"].service_account
           shielded_instance_config = null
+          spot_instance_config     = null
           source_image_family      = null
           source_image_project     = null
           source_image             = null