+++ terraform/examples/slurm_controller_instance/simple/main.tf
@@ -47,6 +47,7 @@ module "slurm_partition0" {
       enable_confidential_vm   = false
       enable_oslogin           = true
       enable_shielded_vm       = false
+      enable_spot_vm           = false
       gpu                      = null
       instance_template        = null
       labels                   = {}
@@ -57,6 +58,7 @@ module "slurm_partition0" {
       preemptible              = false
       service_account          = null
       shielded_instance_config = null
+      spot_instance_config     = null
       source_image_family      = null
       source_image_project     = null
       source_image             = null
@@ -93,6 +95,7 @@ module "slurm_partition1" {
       enable_confidential_vm = false
       enable_oslogin         = true
       enable_shielded_vm     = false
+      enable_spot_vm         = false
       gpu = {
         count = 1
         type  = "nvidia-tesla-v100"
@@ -106,6 +109,7 @@ module "slurm_partition1" {
       preemptible              = false
       service_account          = null
       shielded_instance_config = null
+      spot_instance_config     = null
       source_image_family      = null
       source_image_project     = null
       source_image             = null